refactor(cli): extract huggingface subcommand from voice (#48)
Move HuggingFace model download logic from `voice add` into a dedicated
`kotoba huggingface` subcommand with `add` and `list` actions.

- New `src/cli/huggingface.rs` with `add` (download) and `list` (show downloaded models)
- Remove `VoiceAction::Add` — `voice` now only handles list/set
- Add `Command::Huggingface` + `HuggingFaceAction` enum
- Route new command in main.rs

Closes #48
